rename: #14 global/error 패키지명 수정
global/error 패키지명을 global/exception 으로 수정한다.
