+++ spadeworker/src/main/java/site/devtown/spadeworker/global/exception/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package site.devtown.spadeworker.global.error;
+package site.devtown.spadeworker.global.exception;
 
 import lombok.AccessLevel;
 import lombok.Getter;
